+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Skills added.
 
 ### Changed
+- Fixed showing character with custom theme.
 
 ## v1.33 Beta (07/01/2021)
 ### Added